remove unused variables, use correct types as args, protect against overflow

+++ pd/src/g_bang.c
@@ -132,7 +132,7 @@ static void bng__motionhook(t_scalehandle *sh,
         {
             if (mouse_y > y2)
                 d = mouse_y - y2;
-            else if (abs(mouse_y - y2) < abs(mouse_x - x2))
+            else if (abs((int)(mouse_y - y2)) < abs((int)(mouse_x - x2)))
                 d = mouse_y - y2;
             else
                 d = mouse_x - x2;
